### 2018-05-11 #12 @whedon
[PRE REVIEW]: The Riffomonas Reproducible Research Tutorial Series
**Submitting author:** @pschloss (<a href="http://orcid.org/0000-0002-6935-4275">Patrick D Schloss</a>)
**Repository:** <a href="https://github.com/riffomonas/reproducible_research" target ="_blank">https://github.com/riffomonas/reproducible_research</a>
**Version:** v1.0.0
