### 2018-05-11 #12 @pschloss
Sorry, I'm not sure why the PDF couldn't be built - let me know if I can help debug what went wrong.
