### 2018-05-11 #12 @jhollist
When JOSE is ready for reviewers, would be happy to review this one!
