### 2018-05-14 #12 @labarba
Thank you for agreeing to review, @jhollist ! This gets JOSE started for real.
