+++ _honor_roll/tomkoset.yml
@@ -1,0 +1,8 @@
+---
+name: Sarah Tomkovich
+institution: University of Michigan
+github_handle: tomkoset
+orcid: 0000-0003-1638-5307
+date_completed: 2018-05-18
+layout: card
+---
