### 2018-06-11 #12 @jhollist
All, sorry for the delay on this.  On my list for digging into the review this week.
