### 2018-06-11 #12 @labarba
Good to hear back from you. Have a look at our discussion over at #11, where I'm giving the author more direction about what we want to see in JOSE papers. This, after we had a telecon with the editorial board.

I will start this review now, which will create another issue with the review checklist for you.
