### 2018-06-11 #13 @whedon
[REVIEW]: The Riffomonas Reproducible Research Tutorial Series
**Submitting author:** @pschloss (<a href="http://orcid.org/0000-0002-6935-4275">Patrick D Schloss</a>)
**Repository:** <a href="https://github.com/riffomonas/reproducible_research" target ="_blank">https://github.com/riffomonas/reproducible_research</a>
**Version:** v1.0.0

@jhollist , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:

- [x] **Version:** Does the release version given match the repository release (v1.0.0)?
- [x] **Authorship:** Has the submitting author (@pschloss) made substantial contributions to the module? Does the full list of authors seem appropriate and complete?

- [x] **Example usage:** Do the authors include examples of how to use the module?
