### 2018-07-06 #13 @jhollist
@labarba and @tracykteal, I have a couple of questions about the review checklist.

1. I assume the "statement of need" in **Documentation** and **JOSE Paper** are the same?
2. How should the **Documentation** be presented in the submission?  I am still going through the review but am pretty sure all of the required things will be there.  Didn't know if you were also looking for a more concise doc (i.e. a README) that had all of this in it.

Hope to finish this up today, but if not it'll be week after next (on vacation next week).
