### 2018-07-06 #13 @labarba
It looks like the submitted paper reproduces verbatim the contents of the repository's README. Hmm...
