### 2018-07-06 #13 @labarba
I find it unsatisfying that the paper and the README have the same contents. These documents have different goals., in my opinion

About the paper, the [Author Guide](http://jose.theoj.org/about#author_guidelines) says:

> The goal is that someone reading the JOSE paper has enough information to decide if they'd be interested in adoping the learnig module or software.

The paper is a form of "scholarly advertisement" of the learning module. Readers want to know who the target audience, the design behind the learning module, how it could be used by another instructor or by an independent learner, how it has already been used in the classroom or elsewhere, and of course a description of the contents.

The README is the entry point to the repository, and it might be too lengthy to include all of the above.  It should have installation instructions, list of dependencies, maybe statements about the license and how to contribute,  and include an index of the lessons, with links.
