### 2018-07-06 #13 @jhollist
Thanks for the guidance!  I was leaning in the same direction on the README.  Will include in my review as well.
