### 2018-07-06 #13 @labarba
@pschloss : I would like to see the bullet lists on slides 5 & 6 of the Introduction lesson, just like that, on the paper. It gives the reader an immediate picture of the contents of the module!

Also, in the Statement of Need (in the paper), you give general statements about reproducibility, about the field of microbiome research, and so on, and only in the final sentence refer to this tutorial. What we really want in this section is a statement about _how this tutorial satisfies a particular need_, and perhaps how it differs (or is similar) to other tutorials on the subject.

From the Author Guide:

> Authors make the case for their submission's contribution in the paper, under the heading "Statement of Need."

---

@tracykteal : I'm stepping in here, in this review that you're editing, because with the first handful of papers in JOSE we are in fact defining our genre.
