### 2018-07-06 #13 @pschloss
Thanks for the feedback and my apologies for missing the desired distinction between the paper and the README. Part of the "problem" as you all have identified in the thread here is that it's early days for the journal and I didn't have a lot to go off of for examples.

Should I go ahead and fix things up now as the reviews come in or wait until @jhollist (and @tracykteal?) has had a chance to go through it?
