### 2018-07-09 6356596
First round of revisions
This commit make two changes in response to @labarba in
openjournals/jose-reviews/13

* Revised README to better reflect how to interact with the repository

* Revised paper.md to highlight the statement of need and content that
are covered within the instructional materials
