+++ paper.bib
@@ -93,3 +93,18 @@
   year={2009},
   publisher={Am Soc Microbiol}
 }
+
+@article{Schloss2018,
+  doi = {10.1128/mbio.00525-18},
+  url = {https://doi.org/10.1128/mbio.00525-18},
+  year  = {2018},
+  month = {jun},
+  publisher = {American Society for Microbiology},
+  volume = {9},
+  number = {3},
+  pages = {e00525--18},
+  author = {Patrick D. Schloss},
+  editor = {Jacques Ravel},
+  title = {Identifying and Overcoming Threats to Reproducibility,  Replicability,  Robustness,  and Generalizability in Microbiome Research},
+  journal = {{mBio}}
+}
