+++ paper.md
@@ -26,11 +26,25 @@ The [Riffomonas Reproducible Research tutorial series](https://www.riffomonas.or
 
 
 ## Statement of Need
-A growing number of commentaries speak to a "reproducibility crisis" in science and microbiology in particular [@Collins2014; @Casadevall2016; @Ravel2014; @Garijo2013; @Noble2009]. The relatively new field of microbiome research is interdisciplinary combining microbiology, ecology, bioinformatics, and statistics. Those approaching this field with an interest in soil, aquatic, engineered and host-associated environments must also have background in those areas as well. Although the results of these studies are exciting, the interdisciplinary nature of the field means that the practitioners are not experts in all areas. This is especially true in the analysis of the large datasets that are being generated using high throughput sequencing. The design and structure of the tutorials was motivated by a desire to give scientists new to the field of microbial ecology research the background and tools that they need to implement a reproducible analysis of their data.
+The design and structure of the tutorials in the series give scientists new to the field of microbial ecology research the background and tools that they need to implement a reproducible analysis of their data. Although there are a growing number of commentaries speak to a "reproducibility crisis" in science and microbiology in particular [@Collins2014; @Casadevall2016; @Ravel2014; @Garijo2013; @Noble2009; @Schloss2018], there has not been a concentrated effort to improve reproducibility within microbiology. Although the results of these studies are exciting, the interdisciplinary nature of the field means that the practitioners are not experts in all areas. This is especially true in the analysis of the large datasets that are being generated using high throughput sequencing.
 
 
 ## Content
-The fourteen tutorials focus on issues related to documentation, transparency, openness, and automation. The primer by Noble [@Noble2009] motivates much of the material on project organization, documentation, and automation. Because of the unique nature of computational research relative to bench science, considerable attention is also given to using version control to document the evolution of a project and how to best work individually as well as with others to foster reproducibility. Although the themes are generally tool-agnostic, the practical implementation of these tools is prominent in the materials and when specific tools are used, other options are mentioned. Specific tools that are included in the tutorials include bash for developing command line practices and automation, mothur [@Schloss2009] and R for describing how to script and automate analyses, git and GitHub for introducing version control, make for tracking data dependencies and automation, and the Rmarkdown R package for presenting literate programming techniques. Throughout the tutorial there are exercises that encourage the participants to discuss topics with their mentors and research group, activities where the participants are asked to engage resources at a deeper level. By demonstrating a meaningful level of participation in the tutorial series, participants receive a "virtual badge" and will be listed on the Reproducible Research Tutorial Series Honor Roll, which provides a certification of their training. The material is presented in an escalating manner such that over multiple tutorials participants are introduced to a topic, then given partial exposure and deeper exposure, and finally work with the tools proficiently.
+The fourteen tutorials focus on issues related to documentation, transparency, openness, and automation. The primer by Noble [@Noble2009] motivates much of the material on project organization, documentation, and automation. Because of the unique nature of computational research relative to bench science, considerable attention is also given to using version control to document the evolution of a project and how to best work individually as well as with others to foster reproducibility. Although the themes are generally tool-agnostic, the practical implementation of these tools is prominent in the materials and when specific tools are used, other options are mentioned. Specific tools that are included in the tutorials include:
+
+* Documentation (markdown, Rmarkdown, R, make, git)
+* Organization (bash, HPC/AWS, git)
+* Automation (bash, R, make)
+* Transparency (ORCID, FigShare, git, GitHub, open source licensing)
+* Collaboration (git, GitHub, open source licensing)
+
[…]
