+++ README.md
@@ -1,136 +1,42 @@
-The Riffomonas Reproducible Research Tutorial Series
+README
 =======
 
-Summary
--------
+## Background
+This is a series of tutorials on improving the reproducibility of data analysis for those doing microbial ecology research. Although the materials focus on issues in microbial ecology, the principles are broadly applicable. Also, this series of tutorials is not designed to teach you R or mothur. Again, although the tutorials use R and mothur, you could use other tools (e.g. Python, QIIME) to achieve the same goals. This workshop will focus on the importance of command line practices (e.g. bash), scripting languages (e.g. mothur, R), version control (e.g. git), automation (e.g. make), and literate programming (e.g. Rmarkdown). These are the tools that are used in the Schloss lab to help improve the reproducibility of our manuscripts. By completing the activities in the tutorials you will be listed on the [Reproducible Research Tutorial Honor Roll](https://www.riffomonas.org/reproducible_research/honor_roll), which provides a certification of your training. Accompanying the slides are YouTube-hosted videos where the author presents the material and does live-coding demonstrations to help participants engage with the material. In addition to links to the videos on the tutorial home page, there is a YouTube playlist that aggregates the 14 tutorials as the [Riffomonas Reproducible Research Tutorial
+Series](https://www.youtube.com/watch?v=CfO_f6a3XSo&list=PLmNrK_nkqBpL0d2E26TqPkmTAfelYKbQX). Transcripts from these videos are available as presenter notes for each of the slide decks by pressing 'p' when the slides are open.
 
-The [Riffomonas Reproducible Research tutorial
-series](https://www.riffomonas.org/reproducible_research) is a
-collection of tutorials that focuses on the improvement of reproducible
-data analysis for those doing microbial ecology research. Although the
-materials focus on issues in microbial ecology, the principles are
-broadly applicable. Each tutorial presents broad concepts and how they
-are related to reproducibility as well as applied practice using
-specific tools that are designed to foster reproducibility. In addition
-to activities that are specific to individual tutorials, the series also
-uses data from Kozich et al. (Kozich et al. 2013) to create an example
-that is developed over the course of fourteen tutorials. The tutorial
-series consists of 14 lessons that are available as HTML-based
-presentations with embedded presenter notes taken from live
-demonstrations of the content that are available as videos hosted on
-YouTube.
 
-Statement of Need
------------------
+## Dependencies to build site
+* [Jekyll](https://jekyllrb.com) v.3.3.1
+* [remarkjs](https://remarkjs.com/downloads/remark-latest.min.js) (link to latest version embedded in `resources/_layouts/presentations.html`)
+* [git](https://git-scm.com) v.2.18.0
 
-A growing number of commentaries speak to a "reproducibility crisis" in
-science and microbiology in particular (Collins and Tabak 2014;
-Casadevall et al. 2016; Ravel and Wommack 2014; Garijo et al. 2013;
-Noble 2009). The relatively new field of microbiome research is
-interdisciplinary combining microbiology, ecology, bioinformatics, and
-statistics. Those approaching this field with an interest in soil,
-aquatic, engineered and host-associated environments must also have
-background in those areas as well. Although the results of these studies
-are exciting, the interdisciplinary nature of the field means that the
-practitioners are not experts in all areas. This is especially true in
-the analysis of the large datasets that are being generated using high
-throughput sequencing. The design and structure of the tutorials was
-motivated by a desire to give scientists new to the field of microbial
[…]
