### 2018-07-09 #13 @pschloss
FWIW, I've updated the README and paper.md files to reflect @labarba's comments
