### 2018-07-25 riffomonas/reproducible_research#12 @jhollist
jhollist JOSE Review Issue
## @jhollist review notes

These are my review notes for the JOSE submission at <https://github.com/openjournals/jose-reviews/issues/13>.

This submission introduces a series of tutorials for introducing and practicing tools for reproducible research in microbial ecology. The Riffomonas Reproducible Research Tutorial series is presented as a series of 14 lessons on different reproducibility topics. The lessons are provided either through YouTube Videos or stand alone slide decks with embedded transcripts.  My first impressions of this series is very positive.  The overal quality of the materials, the writing, the content and the presentation on the website is all quite good.  A few things I really like are:
  - the Honor Roll for those that have completed the series
  - the website is laid out well, and easy to read
  - including both slides with transcripts and videos as it allows those with different learning styles to choose the media that works best.
  - focus on reproducbility as a positive and not a crisis
  - highlighting that reproducibility is a spectrum

While my overal impression of this work is positive, I do have one major critique.  The amount of material presented in many of the tutorials can seem overwheleming.  Some of the tutorials have just the right amount of information (e.g. 3. First steps toward reproducibility) while others (e.g. 6. Organization skills for reproducibility) have a LOT of information in them and as organized felt daunting.  I do think a re-organization with links to smaller portions of these larger topics would help quite a bit.  The goal of this kind of re-organization would be to provide users with smaller tutorials that they could conceivably complete within 15-30 minutes of on-line work.

These general comments, as well as a few other items, are captured in the following list.  I am happy to discuss any of the comments, so feel free to follow up in this issue.

### Specific Comments
- [x] General Checks, Version: Could not find an explicitly noted version for this module.
- [x] Documentation, Installation instructions: I do like that the installation instructions for a local build are included as are a list of needed dependencies for the build, but as I read this that isn't necessary to use the tutorials becuase the materials are also available via the website.  Perhaps a first paragraph in the "Installation insturctions" section suggesting that most users of this tutorial can access the materials via the website at <https://www.riffomonas.org/reproducible_research> would help. See also my suggestion below for the Example Usage.
- [x] Documentation, Installation instructions: With this module you have two sets of dependencies, one set for building the Jekyll website and another for working through the R and Reproducible Research (RR) Examples.  I think it is appropriate to talk in detail about the Jekyll Build in the README and have the details of the R and RR dependencies in the module itself.  As is the reader needs to dig a bit to find the dependencies in the module.  In the "Dependencies to run tutorials" section of the README, include a link to the basic dependencies needed for the tutorials.
- [x] Documentation, Example usage: As I was going through this the first time it took me a while to completely grok how I was supposed to interact with the materials.  At first I thought it was the slides, then the videos, then I realized that the slides have the transcript embedded, but that didn't come up until the last link of the Paper or several minutes into the first video.  I think including an "Example Usage", "Quick Start", or "Tutorial Workflow" section that lays out, at a high level, how you would expect users of your tutorial to interact with the content would really help.  Include here your target audience (and expected audiences experience level) and how long it should take users to complete the series.  I think this makes most sense on the website itself, and not in the README.
- [x] Pedagogy / Instructional design, Content scope and length: As I was working on the review I was conflicted about whehter or not using HPC/AWS or local resources is best.  After thinking on it some, I feel it is mostly a design choice and as such is acceptable to keep the focus on HPC/AWS implementations.  That being said, I would like to see some additional information included about running these same tutorials using local resources.  Main reason is that as is, you are expecting users to incur some cost to use the HPC/AWS resources when they could still learn most of the materials with a focus on local tools.  Additionally, both approaches are non-trivial to get up and running and having some instruction for both would improve the tutorial and make it accesible for those that choose to use HPC/AWS or not.
- [x] Pedagogy / Instructional design, Pedagogy: My biggest critique here is that many of the individual modules (e.g. 6, 7) feel big and introduce quite a bit of material for one module.  I don't know that a full re-structuring is required though as the modules themselves could be broken up into more digestable bits.  For instance, the HPC module has 60 slides.  But the first 11 are mostly introduction, slides 12-23 are about working with the cluster, 24-38 are about AWS, 39-54 are Filezilla, 55 - 60 are combining it all together with exercises.  If you have these as 4a, 4b, 4c, and 4d. It lets learners have an obvious place to break.  Could update that navigation of the Tutorials section to reflect this for each of the Tutorials.  For example,  I think 1-3 are good as is, suggestions for 4 outlined above, 6 feels fine up to slide 15, but the Kozich exercise feels a bit out of place here, especially since Git is not covered until the next lesson.  Might make more sense introducing this exercise on it's own?  In short, re-think how each of these tutorials are organized and try to include logical breaks so that the materials are more bite sized.
- [x] paper.md, content section: include citations for R and packages (e.g. rmarkdown)
- [x] paper.md, content Section: When referring to the package, use `rmarkdown`

### Some small things I noticed

- [x] Documentation deck, slide 24: typo - "coul" should be "could"
- [x] Version control deck, slide 42: Missing a trailing ``` on third bullet?
