### 2018-07-25 #13 @jhollist
All, almost done with my review (better late than never and apologies for taking so long).  I have updated the checklist above and am providing a more detailed review in the target repo: <https://github.com/riffomonas/reproducible_research/issues/12>
