### 2018-07-26 #13 @jhollist
Was happy to do this review, @tracykteal and happy to be asked so early in the life of JOSE.  I have some follow up on breaking the tutorials up but will put that over on https://github.com/riffomonas/reproducible_research/issues/12
