### 2018-07-26 riffomonas/reproducible_research#12 @jhollist
Just wanted to follow up on my suggestion to break up some of the tutorials into shorter chunks.

As I was thinking about that I did realize that shooting new videos was likely a non-starter.  A middle ground that I think would help a lot is to keep the organization the change, but provide links to the starting slides for a sections as well as a link to the starting time in the YouTube video that corresponds with that section.

Thoughts?
