### 2018-08-14 riffomonas/reproducible_research#12 @pschloss
Thanks for the feedback @jhollist and @tracykteal. I've been out most of the last month and will be getting back to making revisions this week. Because we're all new to this I have a couple of questions...

* Should I create a separate comment for each of the bullet points listed above, one massive comment, separate comments?
* For the idea of providing links to sections within the slide deck and video would you want to see that at the beginning of each slide deck or within the "Tutorials" outline at http://www.riffomonas.org/reproducible_research/?
