### 2018-08-14 a56ce2d
Add additional references and links for tools mentioned in paper.md, issue #19
