+++ paper.bib
@@ -1,3 +1,20 @@
+@article{rmarkdown,
+    title = {rmarkdown: Dynamic Documents for R},
+    author = {JJ Allaire and Yihui Xie and Jonathan McPherson and Javier Luraschi and Kevin Ushey and Aron Atkins and Hadley Wickham and Joe Cheng and Winston Chang},
+    year = {2018},
+    note = {R package version 1.10},
+    url = {https://CRAN.R-project.org/package=rmarkdown},
+  }
+
+@Manual{R,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2018},
+  url = {https://www.R-project.org/},
+}
+
 @article{Casadevall2016,
   doi = {10.1128/mbio.01256-16},
   url = {https://doi.org/10.1128/mbio.01256-16},
