### 2018-08-14 7c369b9
Improve instructions for how a user would interact with materials
Added a paragraph to `index.md`that does a better job of orienting users
to how they would interact with the tutorial series. Much of this was
already in `paper.md`. Also, restructured and expanded `README.md` to
make it more clear how a typical user versus someone wanting to install
the site locally would interact with the repository. Addresses issue #16
