### 2018-08-14 2e90ebe
Create break points within each tutorial to segment material
