+++ index.md
@@ -34,6 +34,24 @@ title: Reproducible Research
 					<a href="hpc">Using high performance computers</a>
 					<a href="https://youtu.be/5RgRS5VPX1g"><i class="fab fa-youtube"></i></a>
 				</li>
+				<ul>
+					<li>
+						<a href="hpc">Introduction</a>
+						<a href="https://youtu.be/5RgRS5VPX1g"><i class="fab fa-youtube"></i></a>
+					</li>
+					<li>
+						<a href="hpc/#14">Using a local high performance computer</a>
+						<a href="https://youtu.be/5RgRS5VPX1g?t=876"><i class="fab fa-youtube"></i></a>
+					</li>
+					<li>
+						<a href="hpc/#24">Using AWS</a>
+						<a href="https://youtu.be/5RgRS5VPX1g?t=936"><i class="fab fa-youtube"></i></a>
+					</li>
+					<li>
+						<a href="hpc/#39">Setting up FileZilla</a>
+						<a href="https://youtu.be/5RgRS5VPX1g?t=2710"><i class="fab fa-youtube"></i></a>
+					</li>
+				</ul>
 
 				<li>
 					<a href="documentation">The importance of documentation</a>
@@ -44,44 +62,163 @@ title: Reproducible Research
 					<a href="organization">Organizational skills for reproducibility</a>
 					<a href="https://youtu.be/kUnDnmBBGuU"><i class="fab fa-youtube"></i></a>
 				</li>
+				<ul>
+					<li>
+						<a href="organization">Introduction</a>
+						<a href="https://youtu.be/kUnDnmBBGuU"><i class="fab fa-youtube"></i></a>
+					</li>
+					<li>
+						<a href="organization/#10">Organizing a messy project</a>
+						<a href="https://youtu.be/kUnDnmBBGuU?t=1348"><i class="fab fa-youtube"></i></a>
+					</li>
+					<li>
+						<a href="organization/#15">Setting up a new project with a template</a>
+						<a href="https://youtu.be/kUnDnmBBGuU?t=1855"><i class="fab fa-youtube"></i></a>
+					</li>
+				</ul>
 
 				<li>
 					<a href="version_control">Using version control</a>
 					<a href="https://youtu.be/299Anq-Fc4w"><i class="fab fa-youtube"></i></a>
 				</li>
+				<ul>
+					<li>
+						<a href="version_control">Introduction</a>
+						<a href="https://youtu.be/299Anq-Fc4w"><i class="fab fa-youtube"></i></a>
+					</li>
+					<li>
+						<a href="version_control/#22">Setting up and using git</a>
+						<a href="https://youtu.be/299Anq-Fc4w?t=1090"><i class="fab fa-youtube"></i></a>
+					</li>
+					<li>
+						<a href="version_control/#66">Using git with a microbiome project</a>
+						<a href="https://youtu.be/299Anq-Fc4w?t=3270"><i class="fab fa-youtube"></i></a>
+					</li>
+				</ul>
+
 
 				<li>
 					<a href="automation">Automating reproducible analyses</a>
 					<a href="https://youtu.be/57pDlPCodkc"><i class="fab fa-youtube"></i></a>
 				</li>
+				<ul>
+					<li>
+						<a href="automation">Introduction</a>
+						<a href="https://youtu.be/57pDlPCodkc"><i class="fab fa-youtube"></i></a>
+					</li>
+					<li>
+						<a href="automation/#23">Writing bash scripts</a>
+						<a href="https://youtu.be/57pDlPCodkc?t=748"><i class="fab fa-youtube"></i></a>
+					</li>
+				</ul>
+
 
 				<li>
[…]
