### 2018-08-15 089fbd4
Create chapters for longer tutorials
Added links to points in slide decks and videos that correspond to
different "chapters" within the material. Addresses #18.
