### 2018-08-15 #13 @pschloss
Thanks again for the review @jhollist and @tracykteal - I have addressed the comments over at riffomonas/reproducible_research as separate issues. I'm happy to follow up on any of the comments if you have further suggestions or comments.
