### 2018-08-15 riffomonas/reproducible_research#12 @pschloss
I think I've addressed all of your comments in separate issues and in the code. You should be able to see which commit corresponds to each comment. Let me know if you have any other comments or suggestions and I'll be happy to follow up. Thanks!
