### 2018-08-18 #13 @labarba
@jhollist and @tracykteal — Is this submission ready for acceptance?
