### 2018-08-20 riffomonas/reproducible_research#12 @tracykteal
Thanks for creating the separate issues. Everything seems to be addressed. I had one comment on https://github.com/riffomonas/reproducible_research/issues/16 about the color of the font on the Riffomonas website. I created a separate issue too https://github.com/riffomonas/reproducible_research/issues/21 I'm not sure if that's just personal preference or more general around readability and accessibility though.

It sounds like you'd like to wait until things are resolved for JOSE publication and then make version 1.0. Things are resolved from our perspective, so once you make v 1.0 let us know and we'll get it accepted at JOSE.
