### 2018-08-20 #13 @tracykteal
Issues have been resolved. @pschloss will create version 1.0 and then we'll get it in for acceptance.
