### 2018-08-20 #13 @jhollist
I concur!  The issues have been resolved.  I look forward to recommending this material and think it will be a great first pub for JOSE!
