### 2018-08-24 a2ad7b1
Add justification for AWS, addresses #17
