### 2018-08-24 riffomonas/reproducible_research#12 @pschloss
Thanks @jhollist and @tracykteal! I think all of the issues have now been addressed including the assignment of a version number. Please let me know what I need to do next to have the paper accepted and published. This has been a great experience!
