### 2018-08-27 #13 @tracykteal
Sorry, one more thing. Can you create an archive (on Zenodo, figshare, or other) and post the archive DOI here. Then we link to that DOI for the publication.
