### 2018-08-27 5af9251
Add Zenodo DOI badge
