### 2018-08-27 #13 @labarba
@jhollist Not all review-checklist items are checked, but you commented above that you are satisfied with the revision. Can you complete the checklist?
