### 2018-08-29 #13 @jhollist
On vacation now with very flaky connection!  Will finish this up on my return this Saturday.
