### 2018-08-29 #13 @tracykteal
@jhollist had conveyed in the project repository that those items have been addressed, so I updated the checklist. @labarba this should be officially accepted and published. Thanks!
