### 2018-08-30 #13 @labarba
Yay!  @pschloss —Your JOSE paper is now published and you can add the lovely green badge to your repo!

Tremendous thanks to @jhollist for reviewing and @tracykteal for handling the submission. Your volunteer efforts are helping us trail-blaze in new-wave publishing 🎉
